Fine-Tuning and optimization for challenge Falcon Query Compiler vs Vector-Pipes.
Removed bottlenecks:
(1) in column.hpp. Before: the pipeline start invokes ::produce several times to generate a list of all tuplet ids (for full table scan). After: Generating all tuplet ids (from 0 to #items in column) is done via a specialized function ::produce_tupletid_range(a,b) where a is start id and b is end id. It is required that [a,b) is strictly monotonically continuously increasing sequence from a to b (exclusive). This requirement is always satisfied for the set of (all) tuplet ids of a column where no tuplet is marked as "hidden", "removed" or something like that. The implementation of the specialized ::produce function delegates the effort from the operator implementation to the chunk itself, i.e., the chunk class has now an additional function called iota (cf. std::iota) to realize filling the chunk with increasing numbers. In contrast to the previous proceeding, the chunk is always "filled" after a call to this function. Consequentially, chunks can be directly forwarded to subsequent operators.

(2) Removed an unnecessary copying to a temporal value buffer in "materialize.hpp". Before: a dedicated value buffer was used to be filled with materialized values. The memory in this buffer was copied afterwards to the address of the result set. After: the call for materialization now directly writes into the result set, making the effort to manage the value buffer obsolete.