Renamed "Chunks" into "Batches"
As originally proposed as "Vectorized Execution", a vector-at-a-time processing model operates on a subsets of an array. These subsets are defined by an interval [) and the pointer to the original data array is given through the processing. This is slightly different from our approach: Although we run on subsets of a data collection, we are independent of the underlying implementation, and operator on independent data (that was copied with e.g., block_copy) and which is associated to the original data (by the tuplet ids). Hence, we operator on "micro-batches" rather than on real vectors.